# src/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Agent API
  version: 0.1.0
servers:
  - url: "{api_base_url}"
    variables:
      api_base_url:
        default: ""
paths:
...
  /channels/{id}:
    patch:
      summary: Update a channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "Updated" }
    delete:
      summary: Delete a channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "Deleted" }
  /channels/{id}/test:
    post:
      summary: Send a test via a channel
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: "OK" }
...